+++ benchmarks/run_benchmarks.py
@@ -23,7 +23,9 @@ window_sizes = [10, 100,  1000, 10000]
 extra_args = {
     'quantile': 0.75,
     'fill': 0.0,
-    'alpha': 0.1
+    'alpha': 0.1,
+    'lower': -0.1,
+    'upper': 0.5
 }
 
 def main():