+++ bindings/screamer_bindings.cpp
@@ -2,35 +2,91 @@
 #include <pybind11/numpy.h>
 #include <pybind11/stl.h> // Required for std::optional support
 
+#include "screamer/base.h"
+
+#include "screamer/ffill.h"
+#include "screamer/fillna.h"
+#include "screamer/clip.h"
+
 #include "screamer/lag.h"
 #include "screamer/diff.h"
+#include "screamer/return.h"
+#include "screamer/log_return.h"
+
 #include "screamer/rolling_sum.h"
+#include "screamer/rolling_mean.h"
+#include "screamer/rolling_var.h"
+#include "screamer/rolling_std.h"
 #include "screamer/rolling_skew.h"
 #include "screamer/rolling_kurt.h"
 #include "screamer/rolling_zscore.h"
 #include "screamer/rolling_min.h"
 #include "screamer/rolling_max.h"
 #include "screamer/rolling_median.h"
-#include "screamer/return.h"
-#include "screamer/log_return.h"
-#include "screamer/ew_var.h"
-#include "screamer/ew_std.h"
-
-#include "screamer/base.h"
-#include "screamer/rolling_mean.h"
-#include "screamer/rolling_std.h"
-#include "screamer/rolling_var.h"
-#include "screamer/ffill.h"
-#include "screamer/fillna.h"
 #include "screamer/rolling_quantile_ost_v3.h"
+#include "screamer/rolling_rms.h"
 
 #include "screamer/ew_mean.h"
+#include "screamer/ew_var.h"
+#include "screamer/ew_std.h"
+#include "screamer/ew_zscore.h"
+#include "screamer/ew_skew.h"
+#include "screamer/ew_kurt.h"
+#include "screamer/ew_rms.h"
+
+#include "screamer/transform.h"
+
+template <typename T> T signum(T val) {
+    return (T(0) < val) - (val < T(0));
+}
 
 namespace py = pybind11;
 
 PYBIND11_MODULE(screamer_bindings, m) {
 
     py::class_<screamer::ScreamerBase>(m, "_ScreamerBase");
+
+
+
+     py::class_<screamer::Transform<(double (*)(double)) std::abs>>(m, "Abs")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::abs>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::abs>::reset, "Reset to the initial state.");
+
+
+     py::class_<screamer::Transform<(double (*)(double)) std::log>>(m, "Log")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::log>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::log>::reset, "Reset to the initial state.");
+
+     py::class_<screamer::Transform<(double (*)(double)) std::exp>>(m, "Exp")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::exp>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::exp>::reset, "Reset to the initial state.");
+
+     py::class_<screamer::Transform<(double (*)(double)) std::sqrt>>(m, "Sqrt")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::sqrt>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::sqrt>::reset, "Reset to the initial state.");
+
+     py::class_<screamer::Transform<(double (*)(double)) std::erf>>(m, "Erf")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::erf>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::erf>::reset, "Reset to the initial state.");
+
+     py::class_<screamer::Transform<(double (*)(double)) std::erfc>>(m, "Erfc")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) std::erfc>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) std::erfc>::reset, "Reset to the initial state.");
+
+     py::class_<screamer::Transform<(double (*)(double))signum<double> >>(m, "Sign")
+        .def(py::init<>())
+        .def("__call__", &screamer::Transform<(double (*)(double)) signum<double>>::operator(), py::arg("value"))
+        .def("reset", &screamer::Transform<(double (*)(double)) signum<double>>::reset, "Reset to the initial state.");
+
+
+
+
 
     py::class_<screamer::ScreamerBase::LazyIterator>(m, "_LazyIterator")
         .def("__iter__", &screamer::ScreamerBase::LazyIterator::__iter__, py::return_value_policy::reference_internal)
@@ -40,6 +96,11 @@ PYBIND11_MODULE(screamer_bindings, m) {
         .def(py::init<int>(), py::arg("window_size"))
         .def("__call__", &screamer::RollingMean::operator(), py::arg("value"))
         .def("reset", &screamer::RollingMean::reset, "Reset to the initial state.");
+
+    py::class_<screamer::RollingRms, screamer::ScreamerBase>(m, "RollingRms")
+        .def(py::init<int>(), py::arg("window_size"))
+        .def("__call__", &screamer::RollingRms::operator(), py::arg("value"))
+        .def("reset", &screamer::RollingRms::reset, "Reset to the initial state.");
 
     py::class_<screamer::RollingSum, screamer::ScreamerBase>(m, "RollingSum")
         .def(py::init<int>(), py::arg("window_size"))
@@ -120,46 +181,6 @@ PYBIND11_MODULE(screamer_bindings, m) {
         .def(py::init<int>(), py::arg("window_size"))
         .def("__call__", &screamer::LogReturn::operator(), py::arg("value"))
         .def("reset", &screamer::LogReturn::reset, "Reset to the initial state.");
-/*
-    py::class_<screamer::EWMA>(m, "EWMA")
-        .def(py::init<double>(), py::arg("weight"),
-             "Initialize the exponentially weighted moving average indicator.\n\n"
-             ":param weight: Weight of new values.")
-        .def("__call__", &screamer::EWMA::operator(), py::arg("value"),
-             "Update and return the simple moving average.\n\n"
-             ".. math::\n"
-             "    f(X[i]) = w * X[i] + (1 - w) * X[i-1]\n\n"
-             "This applies an exponentially weighted moving average to the stream of input values `X`.")
-        .def("reset", &screamer::EWMA::reset, "Reset the exponentially weighted moving average indicator to its initial state.")
-        .def("transform", &screamer::EWMA::transform, py::arg("input_array"),
-             "Apply the exponentially weighted moving average transformation to a NumPy array.")
-        .doc() = "Exponentially weighted moving average ofvalues in a sequence.";
-
-    py::class_<screamer::EwStd>(m, "EwStd")
-        .def(py::init<double>(), py::arg("weight"),
-             "Initialize the exponentially weighted standard deviation indicator.\n\n"
-             ":param weight: Weight of new values.")
-        .def("__call__", &screamer::EwStd::operator(), py::arg("value"),
-             "Update and return exponentially weighted standard deviation indicator.")
-        .def("reset", &screamer::EwStd::reset, 
-        "Reset the exponentially weighted moving standard deviation indicator to its initial state.")
-        .def("transform", &screamer::EwStd::transform, py::arg("input_array"),
-             "Apply the exponentially weighted standard deviation transformation to a NumPy array.")
-        .doc() = "Exponentially weighted standard deviation of values in a sequence.";
-
-    py::class_<screamer::EwVar>(m, "EwVar")
-        .def(py::init<double>(), py::arg("weight"),
-             "Initialize the exponentially weighted variance indicator.\n\n"
-             ":param weight: Weight of new values.")
-        .def("__call__", &screamer::EwVar::operator(), py::arg("value"),
-             "Update and return exponentially weighted variance indicator.")
-        .def("reset", &screamer::EwVar::reset, 
-        "Reset the exponentially weighted moving standard deviation indicator to its initial state.")
-        .def("transform", &screamer::EwVar::transform, py::arg("input_array"),
-             "Apply the exponentially weighted variance transformation to a NumPy array.")
-        .doc() = "Exponentially weighted variance of values in a sequence.";
-
-*/
 
      py::class_<screamer::EwMean>(m, "EwMean")
         .def(
@@ -176,5 +197,121 @@ PYBIND11_MODULE(screamer_bindings, m) {
         )
         .def("__call__", &screamer::EwMean::operator(), py::arg("value"))
         .def("reset", &screamer::EwMean::reset, "Reset to the initial state.");
+
+     py::class_<screamer::EwVar>(m, "EwVar")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwVar::operator(), py::arg("value"))
+        .def("reset", &screamer::EwVar::reset, "Reset to the initial state.");
+
+     
+     py::class_<screamer::EwStd>(m, "EwStd")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwStd::operator(), py::arg("value"))
+        .def("reset", &screamer::EwStd::reset, "Reset to the initial state.");
+
+     
+     py::class_<screamer::EwZscore>(m, "EwZscore")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwZscore::operator(), py::arg("value"))
+        .def("reset", &screamer::EwZscore::reset, "Reset to the initial state.");
+
+
+     py::class_<screamer::EwSkew>(m, "EwSkew")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwSkew::operator(), py::arg("value"))
+        .def("reset", &screamer::EwSkew::reset, "Reset to the initial state.");
+
+
+     py::class_<screamer::EwKurt>(m, "EwKurt")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwKurt::operator(), py::arg("value"))
+        .def("reset", &screamer::EwKurt::reset, "Reset to the initial state.");
+
+
+     py::class_<screamer::EwRms>(m, "EwRms")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("com") = std::nullopt,
+          py::arg("span") = std::nullopt,
+          py::arg("halflife") = std::nullopt,
+          py::arg("alpha") = std::nullopt
+        )
+        .def("__call__", &screamer::EwRms::operator(), py::arg("value"))
+        .def("reset", &screamer::EwRms::reset, "Reset to the initial state.");
+
+
+
+     py::class_<screamer::Clip>(m, "Clip")
+        .def(
+          py::init<
+               std::optional<double>,
+               std::optional<double>
+          >(),
+          py::arg("lower") = std::nullopt,
+          py::arg("upper") = std::nullopt
+        )
+        .def("__call__", &screamer::Clip::operator(), py::arg("value"))
+        .def("reset", &screamer::Clip::reset, "Reset to the initial state.");
+
 
 }