+++ include/screamer/ew_rms.h
@@ -1,6 +1,6 @@
 // rolling_mean.h
-#ifndef SCREAMER_EW_STD_H
-#define SCREAMER_EW_STD_H
+#ifndef SCREAMER_EW_RMS_H
+#define SCREAMER_EW_RMS_H
 
 #include <optional>
 #include <stdexcept>
@@ -10,9 +10,9 @@
 
 namespace screamer {
 
-    class EwStd : public ScreamerBase {
+    class EwRms : public ScreamerBase {
     public:
-        explicit EwStd(
+        explicit EwRms(
             std::optional<double> com = std::nullopt,
             std::optional<double> span = std::nullopt,
             std::optional<double> halflife = std::nullopt,
@@ -44,99 +44,63 @@ namespace screamer {
                 throw std::invalid_argument("Alpha must be between 0 and 1 (exclusive)");
             }
             one_minus_alpha_ = 1.0 - alpha_;
-            one_minus_alpha2_ = one_minus_alpha_*one_minus_alpha_;
 
             reset();
         }
 
         void reset() override {
-            sum_x_ = 0.0;
             sum_xx_ = 0.0;
             sum_w_ = 0.0;
-            sum_w2_ = 0.0;
         }
 
         double process_scalar(double newValue) override {
-            sum_x_ *= one_minus_alpha_;
             sum_xx_ *= one_minus_alpha_;
-
             sum_w_ *= one_minus_alpha_;
-            sum_w2_ *= one_minus_alpha2_;
 
-            sum_x_ += newValue;
             sum_xx_ += newValue * newValue;
-
             sum_w_ += 1.0;
-            sum_w2_ += 1.0;
 
-            double n_eff = sum_w_* sum_w_ / sum_w2_; 
     
             // Compute the weighted mean
-            double mean = sum_x_ / sum_w_;
-
-            // Compute the weighted variance
-            double variance = (sum_xx_ / sum_w_) - (mean * mean);
-            variance *= n_eff / (n_eff - 1.0);
-            return std::sqrt(variance);
+            double mean = sum_xx_ / sum_w_;
+            return std::sqrt(mean);
         }
 
         void process_array_no_stride(double* y, const double* x, size_t size) override {
             double one_minus_alpha_ = this->one_minus_alpha_;
-            double sum_x_ = 0.0;
             double sum_xx_ = 0.0;
             double sum_w_ = 0.0;
-            double sum_w2_ = 0.0;
 
             for (size_t i=0; i<size; i++) {
-                sum_x_ *= one_minus_alpha_;
                 sum_xx_ *= one_minus_alpha_;
-
                 sum_w_ *= one_minus_alpha_;
-                sum_w2_ *= one_minus_alpha2_;
 
-                sum_x_ += x[i];
                 sum_xx_ += x[i] * x[i];
-
                 sum_w_ += 1.0;
-                sum_w2_ += 1.0;
                 
-                double n_eff = sum_w_ * sum_w_ / sum_w2_; 
-                double mean = sum_x_ / sum_w_;
-                double variance = (sum_xx_ / sum_w_) - (mean * mean);
-                variance *= n_eff / (n_eff - 1.0);
-                y[i] = std::sqrt(variance);      
+                double mean = sum_xx_ / sum_w_;
+                y[i] = std::sqrt(mean);      
             }
         }
 
         void process_array_stride(double* y, size_t dyi, const double* x, size_t dxi, size_t size) override {
             
             double one_minus_alpha_ = this->one_minus_alpha_;
-            double sum_x_ = 0.0;
             double sum_xx_ = 0.0;
             double sum_w_ = 0.0;
-            double sum_w2_ = 0.0;
             
             size_t xi = 0;
             size_t yi = 0;
 
             for (size_t i=0; i<size; i++) { // start at 1
-                sum_x_ *= one_minus_alpha_;
                 sum_xx_ *= one_minus_alpha_;
-
                 sum_w_ *= one_minus_alpha_;
-                sum_w2_ *= one_minus_alpha2_;
 
-                sum_x_ += x[xi];
                 sum_xx_ += x[xi]*x[xi];
-
                 sum_w_ += 1.0;
-                sum_w2_ += 1.0;
 
-                double n_eff = sum_w_* sum_w_ / sum_w2_; 
-                double mean = sum_x_ / sum_w_;
-                double variance = (sum_xx_ / sum_w_) - (mean * mean);
-                variance *= n_eff / (n_eff - 1.0);
-                y[yi] = std::sqrt(variance);
+                double mean = sum_xx_ / sum_w_;
+                y[yi] = std::sqrt(mean);
                 xi += dxi;
                 yi += dyi;                
             }
@@ -146,14 +110,10 @@ namespace screamer {
         double alpha_;
         
         double one_minus_alpha_;
-        double one_minus_alpha2_;
-        double sum_x_;
         double sum_xx_;
         double sum_w_;
-        double sum_w2_;
-        double w2_;
     };
 
 } // namespace screamer
 
-#endif // SCREAMER_EW_VAR_H
+#endif // SCREAMER_EW_RMW_H