+++ reference_impls/scipy.py
@@ -1,0 +1,8 @@
+import scipy.special
+import scipy.special
+
+def Erf__scipy(array):
+    return scipy.special.erf(array)
+
+def Erfc__scipy(array):
+    return scipy.special.erfc(array)